requirements.txt
fastapi
uvicorn[standard]
langchain
langgraph
transformers
torch
requests
feedparser
pydantic
redis
python-dotenv
